+++ CHANGELOG.md
@@ -13,6 +13,7 @@ New features:
   matches the user setting.
 - Messages from the AI are now rendered using the markdown formatting provides.  This makes them
   easier to read and hides the markdown annotations.
+- Added support for Google Gemini 2.0 pro (experimental).
 
 Internal structure changes: